Move getUsersTablature and getUsersPublicInfo to profileServices as they pertain to the /profile route.
