+++ src/pages/Profile/Profile.js
@@ -2,14 +2,15 @@ import { useContext, useEffect, useState } from "react"
 import { useParams } from "react-router-dom"
 import { UserContext } from '../../App'
 import TablatureCard from "../../components/TablatureCard/TablatureCard"
-import * as tablatureServices from "../../services/tablatureServices"
+import * as profileServices from "../../services/profileServices"
 import * as userUtils from "../../utils/userUtils"
 
 // TODO ---
 // Show loading wheel while user is undefined.
-// I need to get all of the users information (profile name, avatar, etc) and public tabs, not just public tabs.
+// Set tablature state depending on if viewing own profile or another users
 // TODO ---
 
+// GET PROFILE INTO STATE
 const Profile = (props) => {
     const [tablature, setTablature] = useState([])
     const [viewingOtherUsersProfile, setViewingOtherUsersProfile] = useState(null)
@@ -19,9 +20,9 @@ const Profile = (props) => {
 
 
     useEffect( () => {
-        const getUsersTablature = async () => {
+        const getUsersPublicInfo = async () => {
             const idToken = userUtils.getIdTokenFromUser(user);
-            return await tablatureServices.getUsersPublicTablature(id, idToken)
+            return await profileServices.getUsersPublicInfo(id, idToken)
         }
 
         if(user) {
@@ -29,7 +30,7 @@ const Profile = (props) => {
                 setTablature(props.usersTablature)
                 setViewingOtherUsersProfile(false)
             } else {
-                getUsersTablature()
+                getUsersPublicInfo()
                 .then( res => {
                     setTablature(res.usersTablature)
                 })
